+++ build.gradle
@@ -19,6 +19,7 @@ repositories {
 }
 
 dependencies {
+    implementation 'org.springframework.boot:spring-boot-starter-validation:2.6.2'
     implementation 'org.springframework.boot:spring-boot-starter-data-jpa:2.6.2'
 //    implementation 'org.springframework.boot:spring-boot-starter-security:2.6.2'
     implementation 'org.springframework.boot:spring-boot-starter-web:2.6.2'
