+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'todolist'
+rootProject.name = 'schedule'
