+++ src/main/java/com/rdr/schedule/ScheduleApplication.java
@@ -1,0 +1,13 @@
+package com.rdr.schedule;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ScheduleApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ScheduleApplication.class, args);
+    }
+
+}
