+++ src/main/java/com/rdr/schedule/common/dto/BaseResponse.java
@@ -1,0 +1,22 @@
+package com.rdr.schedule.common.dto;
+
+import lombok.Getter;
+
+import java.time.LocalDateTime;
+
+@Getter
+public class BaseResponse<T> {
+    private final LocalDateTime localDateTime;
+    private final ResponseMessage message;
+    private final T result;
+
+    public BaseResponse(ResponseMessage message, T result) {
+        this.localDateTime = LocalDateTime.now();
+        this.message = message;
+        this.result = result;
+    }
+
+    public static <T> BaseResponse<T> of(ResponseMessage msg, T result) {
+        return new BaseResponse<>(msg, result);
+    }
+}
