+++ src/main/java/com/rdr/schedule/config/JpaConfig.java
@@ -1,0 +1,9 @@
+package com.rdr.schedule.config;
+
+import org.springframework.context.annotation.Configuration;
+import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
+
+@Configuration
+@EnableJpaAuditing
+public class JpaConfig {
+}
