+++ src/main/java/com/rdr/schedule/schedule/domain/vo/Status.java
@@ -1,0 +1,19 @@
+package com.rdr.schedule.schedule.domain.vo;
+
+public enum Status {
+    TODO,
+    DOING,
+    DONE,
+    DELETED,
+    ;
+
+    public static Status change(Status status) {
+        if (status.equals(Status.TODO)) {
+            return Status.DOING;
+        }
+        if (status.equals(Status.DOING)) {
+            return Status.DONE;
+        }
+        return status;
+    }
+}
