+++ src/main/java/com/rdr/schedule/schedule/dto/bundle/ScheduleGetCountBundle.java
@@ -1,0 +1,19 @@
+package com.rdr.schedule.schedule.dto.bundle;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+public class ScheduleGetCountBundle {
+    private ScheduleGetCountBundle() {
+    }
+
+    @RequiredArgsConstructor
+    @Getter
+    public static class Response {
+        private final Long totalCount;
+
+        public static Response of(Long totalCount) {
+            return new Response(totalCount);
+        }
+    }
+}
