+++ src/main/java/com/rdr/schedule/schedule/dto/request/ScheduleCreateRequest.java
@@ -1,0 +1,13 @@
+package com.rdr.schedule.schedule.dto.request;
+
+import lombok.Getter;
+
+@Getter
+public class ScheduleCreateRequest {
+    private String author;
+    private String title;
+    private String content;
+
+    private ScheduleCreateRequest() {
+    }
+}
