--- src/main/java/com/rdr/todolist/TodolistApplication.java
@@ -1,13 +1,0 @@
-package com.rdr.todolist;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class TodolistApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(TodolistApplication.class, args);
-    }
-
-}
