--- src/main/java/com/rdr/todolist/common/domain/BaseEntity.java
@@ -1,27 +1,0 @@
-package com.rdr.todolist.common.domain;
-
-import lombok.Getter;
-import org.springframework.data.annotation.CreatedDate;
-import org.springframework.data.annotation.LastModifiedDate;
-import org.springframework.data.jpa.domain.support.AuditingEntityListener;
-
-import javax.persistence.Column;
-import javax.persistence.EntityListeners;
-import javax.persistence.MappedSuperclass;
-import java.time.LocalDateTime;
-
-@Getter
-@EntityListeners(AuditingEntityListener.class)
-@MappedSuperclass
-public abstract class BaseEntity {
-    @CreatedDate
-    @Column(name = "CREATE_DATE", nullable = false)
-    protected LocalDateTime createdDate;
-
-    @LastModifiedDate
-    @Column(name = "MODIFIED_DATE", nullable = false)
-    protected LocalDateTime modifiedDate;
-
-    protected BaseEntity() {
-    }
-}
