--- src/main/java/com/rdr/todolist/config/JpaConfig.java
@@ -1,9 +1,0 @@
-package com.rdr.todolist.config;
-
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.jpa.repository.config.EnableJpaAuditing;
-
-@Configuration
-@EnableJpaAuditing
-public class JpaConfig {
-}
