--- src/main/java/com/rdr/todolist/schedule/dto/request/ScheduleCreateRequest.java
@@ -1,13 +1,0 @@
-package com.rdr.todolist.schedule.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class ScheduleCreateRequest {
-    private String author;
-    private String title;
-    private String content;
-
-    private ScheduleCreateRequest() {
-    }
-}
