--- src/main/java/com/rdr/todolist/schedule/dto/request/ScheduleUpdateRequest.java
@@ -1,13 +1,0 @@
-package com.rdr.todolist.schedule.dto.request;
-
-import lombok.Getter;
-
-@Getter
-public class ScheduleUpdateRequest {
-    private Long updateId;
-    private String title;
-    private String content;
-
-    private ScheduleUpdateRequest() {
-    }
-}
