--- src/main/java/com/rdr/todolist/schedule/dto/response/ScheduleFindAllResponse.java
@@ -1,16 +1,0 @@
-package com.rdr.todolist.schedule.dto.response;
-
-import com.rdr.todolist.schedule.dto.bundle.ScheduleFindBundle;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.RequiredArgsConstructor;
-
-import java.util.List;
-
-@RequiredArgsConstructor
-@Builder
-@Getter
-public class ScheduleFindAllResponse {
-    private final Long totalCount;
-    private final List<ScheduleFindBundle.Response> schedules;
-}
