--- src/main/java/com/rdr/todolist/schedule/infrastructure/ScheduleRepository.java
@@ -1,19 +1,0 @@
-package com.rdr.todolist.schedule.infrastructure;
-
-import com.rdr.todolist.schedule.domain.Schedule;
-import org.springframework.data.jpa.repository.JpaRepository;
-import org.springframework.data.jpa.repository.Query;
-
-import java.util.List;
-import java.util.Optional;
-
-public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
-    @Query(value = "select s from Schedule s where not s.status = 'DELETED'")
-    List<Schedule> findAll();
-
-    @Query(value = "select s from Schedule s where not s.status = 'DELETED' and s.id = :id")
-    Optional<Schedule> findById(Long id);
-
-    @Query(value = "select count(s) from Schedule s where not s.status = 'DELETED'")
-    Long getCount();
-}
