+++ src/test/java/com/rdr/schedule/ScheduleApplicationTests.java
@@ -1,0 +1,13 @@
+package com.rdr.schedule;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+
+@SpringBootTest
+class ScheduleApplicationTests {
+
+    @Test
+    void contextLoads() {
+    }
+
+}
