--- src/test/java/com/rdr/todolist/TodolistApplicationTests.java
@@ -1,13 +1,0 @@
-package com.rdr.todolist;
-
-import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-
-@SpringBootTest
-class TodolistApplicationTests {
-
-    @Test
-    void contextLoads() {
-    }
-
-}
